renamed case notes types to form names
